Figure out how to create a reactive file
After much troubleshooting I have figured out the correct way to build a
reactive file using an html file. This commit uses the example provided
by the framework authors. For now, I am going to commit this as a
working sample and iterate on it with the image manipulation idea.