Restart the project from scratch
When originally envisoning this project I wanted to experiment with
julia for mathematical computations. Using Genie has been interesting,
but it is an additional layer of learning a framework when I am most
interested in core julia. My time is limited, so I am switching to a
notebook focused approach to allow my limited time to be focused on my
objectives and not frameworks.