Restart the project offically using Pluto notebook
This commit introduces a very basic Pluto notebook that will serve as
the starting point. As of now, I have one notebook and a working project
structure. This early commit is mostly to upload an image to the GitHub
repo for embedding into the project. Later commits will begin demoing
filtering techniques on this image.